+++ encodings/fastlanes/src/for/vtable/operations.rs
@@ -4,16 +4,17 @@
 use std::ops::Range;
 
 use vortex_array::vtable::OperationsVTable;
-use vortex_array::{Array, ArrayRef, IntoArray};
+use vortex_array::{ArrayRef, IntoArray};
 use vortex_dtype::match_each_integer_ptype;
 use vortex_error::VortexExpect;
 use vortex_scalar::Scalar;
 
-use crate::{FoRArray, FoRVTable};
+use super::FoRVTable;
+use crate::FoRArray;
 
 impl OperationsVTable<FoRVTable> for FoRVTable {
     fn slice(array: &FoRArray, range: Range<usize>) -> ArrayRef {
-        // SAFETY: just slicing encoded data does not affect FOR
+        // SAFETY: Just slicing encoded data does not affect FOR.
         unsafe {
             FoRArray::new_unchecked(
                 array.encoded().slice(range),
@@ -39,8 +40,8 @@ impl OperationsVTable<FoRVTable> for FoRVTable {
                             .vortex_expect("FoRArray Reference value cannot be null"),
                     )
                 })
-                .map(|v| Scalar::primitive::<P>(v, array.dtype().nullability()))
-                .unwrap_or_else(|| Scalar::null(array.dtype().clone()))
+                .map(|v| Scalar::primitive::<P>(v, array.reference_scalar().dtype().nullability()))
+                .unwrap_or_else(|| Scalar::null(array.reference_scalar().dtype().clone()))
         })
     }
 }
